[type: refactor] refactor exchanger
